# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MiniFlow API
  version: 0.1.0
  description: API contract for MiniFlow MVP.

servers:
  - url: /api/v1

tags:
  - name: Auth
    description: Cookie-based authentication endpoints
  - name: Requests
    description: Request workflow endpoints

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /auth/login:
    post:
      tags: [Auth]
      summary: Login user and set auth cookies
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Auth and CSRF cookies are set by the server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout user and clear cookies
      operationId: logoutUser
      security:
        - cookieAuth: []
        - csrfHeader: []
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      operationId: getCurrentUser
      security:
        - cookieAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/csrf:
    get:
      tags: [Auth]
      summary: Issue CSRF token
      operationId: getCsrfToken
      security:
        - cookieAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CsrfToken"

  /requests:
    post:
      tags: [Requests]
      summary: Create draft request
      operationId: createRequest
      security:
        - cookieAuth: []
        - csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
    get:
      tags: [Requests]
      summary: List requests
      operationId: listRequests
      parameters:
        - name: teamId
          in: query
          required: true
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/RequestStatus"
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: includeDeleted
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListRequestsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /requests/{id}:
    patch:
      tags: [Requests]
      summary: Update request (Draft only)
      operationId: updateRequest
      security:
        - cookieAuth: []
        - csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRequestBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
    get:
      tags: [Requests]
      summary: Get request detail
      operationId: getRequest
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /requests/{id}/submit:
    post:
      tags: [Requests]
      summary: Submit request (Draft -> Pending)
      operationId: submitRequest
      security:
        - cookieAuth: []
        - csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /requests/{id}/approve:
    post:
      tags: [Requests]
      summary: Approve request (Pending -> Approved)
      operationId: approveRequest
      security:
        - cookieAuth: []
        - csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /requests/{id}/reject:
    post:
      tags: [Requests]
      summary: Reject request (Pending -> Rejected)
      operationId: rejectRequest
      security:
        - cookieAuth: []
        - csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /requests/{id}/revise:
    post:
      tags: [Requests]
      summary: Revise request (Rejected -> Draft)
      operationId: reviseRequest
      security:
        - cookieAuth: []
        - csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /requests/{id}/delete:
    post:
      tags: [Requests]
      summary: Logical delete request (Draft/Rejected -> Deleted)
      operationId: deleteRequest
      security:
        - cookieAuth: []
        - csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

components:
  parameters:
    RequestId:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
    XActorId:
      name: x-actor-id
      in: header
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
    XTeamId:
      name: x-team-id
      in: header
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"

  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            badRequest:
              value:
                error:
                  code: INVALID_INPUT
                  message: Invalid input.
                  status: 400
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            forbidden:
              value:
                error:
                  code: FORBIDDEN
                  message: You are not allowed to perform this action.
                  status: 403
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            unauthorized:
              value:
                error:
                  code: UNAUTHORIZED
                  message: authentication required
                  status: 401
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notFound:
              value:
                error:
                  code: NOT_FOUND
                  message: Request not found.
                  status: 404
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            conflict:
              value:
                error:
                  code: STATE_CONFLICT
                  message: Operation is not allowed in current status.
                  status: 409
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internalServerError:
              value:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: Unexpected server error.
                  status: 500

  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: miniflow_auth
    csrfHeader:
      type: apiKey
      in: header
      name: x-csrf-token

  schemas:
    Uuid:
      type: string
      format: uuid

    RequestStatus:
      type: string
      enum: [Draft, Pending, Approved, Rejected, Deleted]

    ApprovalActionType:
      type: string
      enum: [Approved, Rejected]

    Request:
      type: object
      required:
        - id
        - teamId
        - title
        - body
        - status
        - createdBy
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        teamId:
          type: string
        title:
          type: string
          minLength: 1
          maxLength: 200
        body:
          type: string
          minLength: 1
          maxLength: 5000
        status:
          $ref: "#/components/schemas/RequestStatus"
        createdBy:
          $ref: "#/components/schemas/Uuid"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: [string, "null"]
          format: date-time

    Approval:
      type: object
      required:
        - id
        - requestId
        - actedBy
        - actionType
        - createdAt
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        requestId:
          $ref: "#/components/schemas/Uuid"
        actedBy:
          $ref: "#/components/schemas/Uuid"
        actionType:
          $ref: "#/components/schemas/ApprovalActionType"
        reason:
          type: [string, "null"]
          maxLength: 1000
        createdAt:
          type: string
          format: date-time

    RequestDetail:
      type: object
      required:
        - request
        - approvals
      properties:
        request:
          $ref: "#/components/schemas/Request"
        approvals:
          type: array
          items:
            $ref: "#/components/schemas/Approval"

    AuthUser:
      type: object
      required: [id, email]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        email:
          type: string
          format: email

    AuthUserResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8

    RegisterInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8

    CsrfToken:
      type: object
      required: [csrfToken]
      properties:
        csrfToken:
          type: string

    CreateRequestBody:
      type: object
      required: [teamId, title, body]
      properties:
        teamId:
          type: string
          minLength: 1
          maxLength: 200
        title:
          type: string
          minLength: 1
          maxLength: 200
        body:
          type: string
          minLength: 1
          maxLength: 5000

    UpdateRequestBody:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        body:
          type: string
          minLength: 1
          maxLength: 5000

    DecisionBody:
      type: object
      properties:
        reason:
          type: string
          maxLength: 1000

    ErrorObject:
      type: object
      required: [code, message, status]
      properties:
        code:
          type: string
        message:
          type: string
        status:
          type: integer

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorObject"

    ListRequestsResponse:
      type: object
      required: [items, page, limit, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Request"
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0